requirements.txt
streamlit
opencv-python
numpy
easyocr
networkx
requests
python-dotenv
google-auth
pandas
scikit-learn
